# openapi.yaml
openapi: 3.0.3
info:
  title: Games API
  version: 1.0.0
servers:
  - url: http://localhost:3000
tags:
  - name: games
    description: Basic info of your games
components:
  schemas:
    Game:
      required:
        - title
        - developer
        - distributor
        - genre
        - released
      type: object
      properties:
        title:
          type: string
          example: Galaga
        developer:
          type: string
          example: Namco
        distributor:
          type: string
          example: Namco
        genre:
          type: string
          example: Fixed Shooter
        released:
          type: string
          example: "1981-09-25"
paths:
  /games:
    get:
      tags:
        - games
      summary: Find all games
      description: Returns all games
      responses:
        "200":
          description: Returns a JSON containing all registered games
  /games/{gameID}:
    get:
      tags:
        - games
      summary: Find a game by ID
      description: Returns a single game
      parameters:
        - name: gameID
          in: path
          description: ID of game to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Returns a JSON object of the game
    delete:
      tags:
        - games
      summary: Delete a game by ID
      description: Deletes a  game
      parameters:
        - name: gameID
          in: path
          description: ID of game to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Returns a JSON object of the deleted game
    post:
      tags:
        - games
      summary: Add a game
      description: Add a new game to the database
      requestBody:
        description: Create a new game
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
      responses:
        "200":
          description: Returns a JSON object of the game